requirements_prod.txt
boto3==1.9.38
botocore==1.12.38
docutils==0.14
jmespath==0.9.3
numexpr==2.6.8
numpy==1.15.3
pandas==0.23.4
psycopg2-binary
psycopg2==2.7.5
python-dateutil==2.7.5
python-dotenv==0.10.0
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.13
tables==3.4.4
urllib3==1.24.1
Click==7.0
